refactor(shared): separate `postSchema` to `schemas/posts.ts`
